Changeset for multi level job
When we set up a multi level job, the leaf jobs were not able to get the change set. Due to which the game is not able to update the score for the concerned player. Now it would recursively get to the topmost job and add that job to the list of accountable builds.